requirements.txt
 Pillow
 pyproj
 mapnik
 django
